chore(indexer-alt): factor out ingestion clients and errors
